Move UniverseGenerator.{h,cpp} into server directory
