+++ server/ServerApp.cpp
@@ -2,6 +2,7 @@
 
 #include "SaveLoad.h"
 #include "ServerFSM.h"
+#include "UniverseGenerator.h"
 #include "../combat/CombatSystem.h"
 #include "../combat/CombatEvents.h"
 #include "../combat/CombatLogManager.h"
@@ -18,7 +19,6 @@
 #include "../universe/System.h"
 #include "../universe/Species.h"
 #include "../universe/Tech.h"
-#include "../universe/UniverseGenerator.h"
 #include "../universe/UnlockableItem.h"
 #include "../universe/Enums.h"
 #include "../universe/ValueRef.h"
