+++ server/ServerWrapper.cpp
@@ -1,6 +1,7 @@
 #include "ServerWrapper.h"
 
 #include "ServerApp.h"
+#include "UniverseGenerator.h"
 
 #include "../universe/Condition.h"
 #include "../universe/ScriptingContext.h"
@@ -17,7 +18,6 @@
 #include "../universe/Tech.h"
 #include "../universe/Pathfinder.h"
 #include "../universe/Universe.h"
-#include "../universe/UniverseGenerator.h"
 #include "../universe/UnlockableItem.h"
 #include "../universe/Enums.h"
 #include "../universe/ValueRef.h"
